# apps/casino/cmd/api/docs/swagger.yaml
basePath: /
definitions:
  http.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.RegisterReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Бэкенд авторизации для казино-проекта.
  title: Casino API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/http.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход пользователя
      tags:
      - auth
  /api/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Информация о текущем пользователе
      tags:
      - users
  /api/register:
    post:
      consumes:
      - application/json
      description: Создает пользователя в системе и возвращает JWT
      parameters:
      - description: Email и пароль
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/http.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"