# apps/casino/internal/api/dto.go
package api

type RegisterReq struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
}

type LoginReq struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}